\name{EnvObj}
\alias{EnvObj}
\title{Environment's Objects}
\description{
  Get an environment's visible objects as a list.
}
\usage{
  EnvObj(envir = e)
}
\arguments{
  \item{envir}{environment to get its content}
}
\details{
  All the visible objects in the environment including functions and data will be returned.
}
\value{
  All visible objects as a list
}
\author{Kyun-Seop Bae <k@acr.kr>}
